change beacon.level to double, erase not used imports, fix in ESRepository.create now it is using the Entity id

+++ src/main/scala/com/teracode/beacons/domain/Domain.scala
@@ -16,12 +16,12 @@ trait Location {
 
 trait Beacon {
   def ssid: String
-  def level: Int
+  def level: Double
 }
 
 /* Data */
 case class LocationData(name: String, description: String, status: String, signals: List[BeaconData]) extends Location
-case class BeaconData(ssid: String, level: Int) extends Beacon
+case class BeaconData(ssid: String, level: Double) extends Beacon
 
 /* Entity */
 case class LocationEntity(id: UUID, name: String, description: String, status: String, signals: List[BeaconData]) extends Location with Entity
